+++ lib/src/framework/core/dependencies/global_widgets/hover_global_widgets.dart
@@ -4,14 +4,18 @@ import 'package:provider/provider.dart';
 // Global components that will be reused throughout the app.
 // User provider to pass the global elements to HoverPage widgets.
 class HoverGlobalWidgets {
-  HoverGlobalWidgets();
-  Widget appBar;
-  Widget drawer;
-  Widget snackBar;
-  Widget floatingActionButton;
+  HoverGlobalWidgets({
+    this.appBarBuilder,
+    this.drawerBuilder,
+    this.snackBarBuilder,
+    this.floatingActionButtonBuilder,
+  });
+  final Widget Function(BuildContext) appBarBuilder;
+  final Widget Function(BuildContext) drawerBuilder;
+  final Widget Function(BuildContext) snackBarBuilder;
+  final Widget Function(BuildContext) floatingActionButtonBuilder;
 }
 
 class HoverGlobalWidgetsProvider extends Provider<HoverGlobalWidgets> {
-  HoverGlobalWidgetsProvider(HoverGlobalWidgets globalElements)
-      : super(create: (_) => globalElements);
+  HoverGlobalWidgetsProvider(HoverGlobalWidgets globalElements) : super(create: (_) => globalElements);
 }