+++ lib/src/framework/core/dependencies/navigation/routing/hover_route.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/widgets.dart';
 
-abstract class HoverRoute extends Widget {
+abstract class HoverRoute extends StatelessWidget {
   String get routeName;
 }